refactor: 필터링 관련 폴더 구조 변경
모달 필터가 추가 될 예정이기 때문에, 상위 폴더인 Filters 생성하고 SideFilter를 하위에 위치시킴
