+++ frontend/src/components/Filters/RangeBar/index.tsx
