+++ frontend/src/components/Filters/RangeBar/style.ts
